Add New Challan, View Challan Item Done

+++ src/main/java/com/bohniman/vmsmaintenance/model/TransItemPurchase.java
@@ -28,7 +28,7 @@ public class TransItemPurchase {
 
     private Date purchaseDate;
 
-    private Long puchaseQuantity;
+    private Double puchaseQuantity;
 
     // private Long amount;
